% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BoxCox}
\alias{BoxCox}
\title{One-parameter Box-Cox transformation.}
\usage{
BoxCox(x, lambda)
}
\arguments{
\item{x}{(numeric) data to be transformed}

\item{lambda}{(numeric) Box-Cox transformation parameter}
}
\value{
(numeric) vector with Box-Cox transformation of x
}
\description{
One-parameter Box-Cox transformation.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
