% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{computePerfMeasAll}
\alias{computePerfMeasAll}
\title{Function for computing reference intervals for all markers}
\usage{
computePerfMeasAll(analytes, algo, risIn, tableTCs, cutoffZ = 5)
}
\arguments{
\item{analytes}{(character) listing all analytes for which the result files should be parsed}

\item{algo}{(character)    specifying used algorithm}

\item{risIn}{(list) with data frame of all calculated reference intervals}

\item{tableTCs}{(data.frame) containing all information about the simulated test sets}

\item{cutoffZ}{(integer) specifying if and if so which cutoff should be used to classify results as implausible (default: 5)}
}
\value{
list with the calculated errors as data frame for each marker
}
\description{
Function for computing reference intervals for all markers
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
