% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{computeRIs}
\alias{computeRIs}
\title{Function for computing reference intervals}
\usage{
computeRIs(
  analyte,
  algo,
  results,
  tableTCs,
  RIperc = c(0.025, 0.975),
  truncNormal = FALSE
)
}
\arguments{
\item{analyte}{(character) specifiyng analyte}

\item{algo}{(character)    specifying used algorithm}

\item{results}{(list) with all calculated results as RWDRI objects}

\item{tableTCs}{(data frame) containing all information about the simulated test sets}

\item{RIperc}{(numeric) vector specifying the percentiles for the reference interval, default: 0.025 and 0.975}

\item{truncNormal}{(logical) specifying if a normal distribution truncated at zero shall be assumed}
}
\value{
data frame with computed reference intervals
}
\description{
Function for computing reference intervals
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
