% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{computeRIsAll}
\alias{computeRIsAll}
\title{Function for computing reference intervals for all markers}
\usage{
computeRIsAll(analytes, algo, resIn, tableTCs, truncNormal = FALSE)
}
\arguments{
\item{analytes}{(character) listing all markers for which the result files should be parsed}

\item{algo}{(character)    specifying used algorithm}

\item{resIn}{(list) with all calculated results for all markers as RWDRI objects}

\item{tableTCs}{(data.frame) containing all information about the simulated test sets}

\item{truncNormal}{(logical) specifying if a normal distribution truncated at zero shall be assumed}
}
\value{
list with the calculated reference intervals as data frame for each marker
}
\description{
Function for computing reference intervals for all markers
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
