% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{computeRuntimeAll}
\alias{computeRuntimeAll}
\title{Function to compute runtime statistics for all analytes}
\usage{
computeRuntimeAll(analytes, algo, risIn, tableTCs)
}
\arguments{
\item{analytes}{(character) listing all analytes for which the result files should be parsed}

\item{algo}{(character)    specifying used algorithm}

\item{risIn}{(list) with data frame of all calculated reference intervals and runtime}

\item{tableTCs}{(data.frame) containing all information about the simulated test cases}
}
\value{
(list) wit runtime statistics per analyte and data frames with raw runtime overall and per analyte
}
\description{
Function to compute runtime statistics for all analytes
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
