% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{computeSubResults}
\alias{computeSubResults}
\title{Helper function to compute the subscores for the distribution types and the mentioned categories}
\usage{
computeSubResults(
  errorDf,
  tableTCs,
  distCat,
  errorParam,
  catList,
  catLabels,
  perfCombination = "mean"
)
}
\arguments{
\item{errorDf}{(data frame) containing the estimate reference intervals and all computed error measures}

\item{tableTCs}{(data.frame) containing all information about the simulated test sets}

\item{distCat}{(character) specifying the distribution category}

\item{errorParam}{(character) specifiying for which error parameter the data frame should be generated}

\item{catList}{(character) vector containing the categories to split the dataset}

\item{catLabels}{(character) vector containing the labels that will be used for the categories}

\item{perfCombination}{(character) specifying if mean (default), median or sum should be computed}
}
\value{
(data frame) containing the computed subscores
}
\description{
Helper function to compute the subscores for the distribution types and the mentioned categories
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
