% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defineSubset}
\alias{defineSubset}
\title{Function for defining a subset that is used for analyizing the computation time and can be used for other subanalyses.}
\usage{
defineSubset(tableTCs = NULL, N = 50, seed = 123)
}
\arguments{
\item{tableTCs}{(data frame) describing the pre-defined testcases}

\item{N}{(integer) describing the number of testcases per biomarker contained in the subset (default: 50)}

\item{seed}{(integer) specifying the seed used for defining the subset, default: 123}
}
\value{
(data frame) describing the updated table with all test case definitions.
}
\description{
Function for defining a subset that is used for analyizing the computation time and can be used for other subanalyses.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
