% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatNumber}
\alias{formatNumber}
\title{Rounding method with trailing zeros.}
\usage{
formatNumber(x, digits)
}
\arguments{
\item{x}{(numeric) value that is rounded}

\item{digits}{(integer) indicating the number of decimal places to be used}
}
\value{
Rounded value with trailing zeros
}
\description{
Rounding method with trailing zeros.
}
\author{
Christopher Rank \email{christopher.rank@roche.com}
}
