% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateMD5}
\alias{generateMD5}
\title{Generate an MD5 hash sum for any R object.}
\usage{
generateMD5(x)
}
\arguments{
\item{x}{(object) any R object.}
}
\value{
(character) MD5 hash sum of the input object.
}
\description{
Generate an MD5 hash sum for any R object.
}
\author{
Christopher Rank \email{christopher.rank@roche.com}
}
