% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{getRIsAllwithoutModel}
\alias{getRIsAllwithoutModel}
\title{Function for retrieving reference intervals if directly computed}
\usage{
getRIsAllwithoutModel(analytes, algo, resIn, tableTCs)
}
\arguments{
\item{analytes}{(character) listing all markers for which the result files should be parsed}

\item{algo}{(character)    specifying used algorithm}

\item{resIn}{(list) with all calculated results for all markers as RWDRI objects}

\item{tableTCs}{(data.frame) containing all information about the simulated test sets}
}
\value{
list with the calculated reference intervals as data frame for each marker
}
\description{
Function for retrieving reference intervals if directly computed
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
