% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invBoxCox}
\alias{invBoxCox}
\title{Inverse of the one-parameter Box-Cox transformation.}
\usage{
invBoxCox(x, lambda)
}
\arguments{
\item{x}{(numeric) data to be transformed}

\item{lambda}{(numeric) Box-Cox transformation parameter}
}
\value{
(numeric) vector with inverse Box-Cox transformation of x
}
\description{
Inverse of the one-parameter Box-Cox transformation.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
