% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadTestsetDefinition}
\alias{loadTestsetDefinition}
\title{Convenience function to load the table with the information about the pre-defined test sets}
\usage{
loadTestsetDefinition()
}
\value{
(data frame) containing the pre-defined parameter combinations to generate the simulations
}
\description{
Convenience function to load the table with the information about the pre-defined test sets
}
\examples{
testsets <- loadTestsetDefinition()
str(testsets)

}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
