% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{mergeSummaryErrors}
\alias{mergeSummaryErrors}
\title{Helper function to combine all computed summary errors}
\usage{
mergeSummaryErrors(
  errorList,
  nameVec,
  errorParam = "MedianAbsPercErrorOV",
  cutoffZ = FALSE
)
}
\arguments{
\item{errorList}{(list) of the error lists for the different methods for which the summary errors should be combined}

\item{nameVec}{(character) vector specifying the names of the methods}

\item{errorParam}{(character) specifying for which error parameter the data frame should be generated}

\item{cutoffZ}{(logical) indicating if a cutoff was set, needed for CRP case}
}
\value{
(data frame) containing the summary errors per analyte per method
}
\description{
Helper function to combine all computed summary errors
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
