% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plot method forrro generating a boxplot of the benchmark results}
\usage{
plotBoxplot(
  errorList,
  colList,
  nameList,
  outline = TRUE,
  withMean = TRUE,
  withCats = FALSE,
  withDirect = TRUE,
  title = "",
  outputDir = NULL,
  filename = NULL,
  ylim1 = c(0, 100),
  ylim2 = c(100, 1000),
  ...
)
}
\arguments{
\item{errorList}{containing the overall benchmark results}

\item{colList}{(character) vector specifying the colors used for the different algorithms (should correspond to columns of benchmark results)}

\item{nameList}{(character) vector specifying the names used in the legend (should correspond to columns of benchmark results), if NULL, colnames will be used}

\item{outline}{(logical) indicating whether outliers should be drawn (TRUE, default), or not (FALSE)}

\item{withMean}{(logical) indicating whether the mean should be plotted as well (default: TRUE)}

\item{withCats}{(logical) set to TRUE if categories (e.g. pathological fraction) should be plotted (default: FALSE)}

\item{withDirect}{(logical) indicating whether the box of the direct method should be elongated to facilitate comparison (default:TRUE)}

\item{title}{(character) specifying plot title}

\item{outputDir}{(character) specifying a output directory}

\item{filename}{(character) specifying a filename for the plot}

\item{ylim1}{(numeric) vector specifying the limits in y-direction for the first granular scale}

\item{ylim2}{(numeric) vector specifying the limits in y-direction for the second less detailed scale}

\item{...}{additional arguments passed forward to other functions}
}
\value{
No return value. Instead, a plot is generated.
}
\description{
Plot method forrro generating a boxplot of the benchmark results
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
