% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotScatterplot}
\alias{plotScatterplot}
\title{Plot method for generating a scatterplot}
\usage{
plotScatterplot(
  errorList,
  colList,
  nameList,
  withColor = NULL,
  cats = NULL,
  title = "",
  outputDir = NULL,
  filename = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{errorList}{(data frame) containing the overall benchmark results}

\item{colList}{(character) vector specifying the colors used for the different algorithms (should correspond to columns of benchmark results)}

\item{nameList}{(character) vector specifying the names used in the legend (should correspond to columns of benchmark results), if NULL, colnames will be used}

\item{withColor}{(character) indicating if plot should be colored according to pathological fraction, sample size or pathological overlap left / right}

\item{cats}{(character) specifying the category labels}

\item{title}{(character) specifying plot title}

\item{outputDir}{(character) specifying a output directory}

\item{filename}{(character) specifying a filename for the plot}

\item{xlim}{(numeric) vector specifying the limits in y-direction for the first granular scale}

\item{ylim}{(numeric) vector specifying the limits in y-direction for the second less detailed scale}

\item{xlab}{(character) specifying x-axis label}

\item{ylab}{(character) specifying y-axis label}

\item{...}{additional arguments passed forward to other functions}
}
\value{
No return value. Instead, a plot is generated.
}
\description{
Plot method for generating a scatterplot
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
