% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressInd}
\alias{progressInd}
\title{Function for setting up the progress indicator.}
\usage{
progressInd(value, maxValue, nCharMsg = 0)
}
\arguments{
\item{value}{(integer) indicating the current number}

\item{maxValue}{(integer) indicating the maximum number}

\item{nCharMsg}{(integer) indicating the number of characters the message already has}
}
\value{
(character) returing generated progress message
}
\description{
Function for setting up the progress indicator.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
