% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{readResultFiles}
\alias{readResultFiles}
\title{Function for reading in the result files for one marker}
\usage{
readResultFiles(analyte, algo, path = NULL, tableTCs = NULL)
}
\arguments{
\item{analyte}{(character) specifying analyte}

\item{algo}{(character)    specifying used algorithm}

\item{path}{(character) specifying path to Results directories}

\item{tableTCs}{(data frame) containing all information about the simulated test sets}
}
\value{
list with caluclated results as RWDRI objects
}
\description{
Function for reading in the result files for one marker
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
