% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{readResultFilesAll}
\alias{readResultFilesAll}
\title{Function for reading all results files.}
\usage{
readResultFilesAll(analytes, algo, baseDir = NULL, inputDir = NULL, tableTCs)
}
\arguments{
\item{analytes}{(character) listing all analytes for which the result files should be parsed}

\item{algo}{(character)    specifying used algorithm}

\item{baseDir}{(character) specifying the baseDir: Results will be used from baseDir/Results/algo/marker 
if baseDir is set, inputDir will be ignored; if baseDir is NULL, the current working directory will be used}

\item{inputDir}{(character) specifying path directly to Results directories}

\item{tableTCs}{(data frame) containing all information about the simulated test sets}
}
\value{
list with all caluclated results as RWDRI objects for each marker
}
\description{
Function for reading all results files.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
