% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateResults.R
\name{restrictSet}
\alias{restrictSet}
\title{Function to get error subsets for defined category and restriction.}
\usage{
restrictSet(overallCat, tableTCs, errorList, distType = TRUE, restrict = NULL)
}
\arguments{
\item{overallCat}{(list) containing the categories to split the dataset}

\item{tableTCs}{(data.frame) containing all information about the simulated test sets}

\item{errorList}{(list) containing for each method the table with the computed error measurements}

\item{distType}{(logical) indicating if 'overallCat' refers to analytes (FALSE, default) or distribution types (TRUE)}

\item{restrict}{(character) indicating whether testcases should be filtered according to specified restriction, default NULL, 
e.g. fractionPathol <= 0.30}
}
\value{
(list) containing the merged performance measurements grouped according to specified category
}
\description{
Function to get error subsets for defined category and restriction.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
