% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupDirStructure}
\alias{setupDirStructure}
\title{Convenience function to set up the directory structure used for storing data and results.}
\usage{
setupDirStructure(
  outputDir = NULL,
  onlyData = FALSE,
  onlyResults = FALSE,
  tableTCs = NULL,
  algoName = NULL
)
}
\arguments{
\item{outputDir}{(character) specifying the base output directory. From here, Data/biomarker and Result/algoName/biomarker directories are generated}

\item{onlyData}{(logical) if set to TRUE, only the biomarker subdirectories are generated, name of output directory is used as it is (default:FALSE)}

\item{onlyResults}{(logical) if set to TRUE, only the algoName/biomarker subdirectories are generated, name of output directory is used as it is (default:FALSE)}

\item{tableTCs}{(data frame) containing the pre-defined parameter combinations to generate the simulations}

\item{algoName}{(character) specifying the name of the algorithm used for creating the subdirectory}
}
\value{
No return value. Instead, the directory structure is set up.
}
\description{
Convenience function to set up the directory structure used for storing data and results.
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
