% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateTestSets.R
\name{writeResFile}
\alias{writeResFile}
\title{Helper function to write result file when time out occured or R session terminated}
\usage{
writeResFile(
  algoName,
  biomarker,
  N = 0,
  error = NULL,
  runtime = NULL,
  filename = NULL,
  outputDir = NULL
)
}
\arguments{
\item{algoName}{(character) specifying the name of the algorithm that is evaluated}

\item{biomarker}{(character) specifying the biomarker for which the algorithm should calculate RIs}

\item{N}{(numeric) specifying the number of input data points}

\item{error}{(character) specifying the type of error (e.g. timeout, RSessionTerminated)}

\item{runtime}{(numeric) specifying the computation time up until the error occured}

\item{filename}{(character) specifying the filename for which the algorithm failed}

\item{outputDir}{(character) specifying the outputDir: Data files should be stored in outputDir/Data/biomarker and Results will be stored in outputDir/Results/algo/biomarker}
}
\description{
Helper function to write result file when time out occured or R session terminated
}
\author{
Tatjana Ammer \email{tatjana.ammer@roche.com}
}
